fix: support object type for package.json workspaces field
