+++ packages/modular-scripts/src/addPackage.ts
@@ -230,8 +230,8 @@ async function addPackage({
       modular: {
         type: modularTemplateType,
       },
-      main: modularTemplatePackageJson.main,
-      dependencies: modularTemplatePackageJson.dependencies,
+      main: modularTemplatePackageJson?.main,
+      dependencies: modularTemplatePackageJson?.dependencies,
       version: '1.0.0',
     },
     {
