+++ packages/modular-scripts/src/utils/isModularType.ts
@@ -1,5 +1,7 @@
-import type { JSONSchemaForNPMPackageJsonFiles as PackageJson } from '@schemastore/package';
 import * as path from 'path';
+
+import type { JSONSchemaForNPMPackageJsonFiles as PackageJson } from '@schemastore/package';
+
 import * as fs from 'fs-extra';
 
 export const packageTypes: PackageType[] = [
@@ -19,13 +21,30 @@ export type PackageType = ModularTemplateType | 'template';
 
 export type ModularType = PackageType | 'root';
 
-export type ModularPackageJson = PackageJson & {
+// Utility type that extends type `T1` with the fields of type `T2`
+type Extend<T1, T2> = {
+  [k in keyof (T1 & T2)]: k extends keyof T2
+    ? T2[k]
+    : k extends keyof T1
+    ? T1[k]
+    : never;
+};
+
+type PackageJsonOverrides = {
   browserslist?: Record<string, string[]>;
   modular?: {
     type: ModularType;
     templateType?: ModularTemplateType;
   };
+  workspaces?:
+    | string[]
+    | {
+        packages?: string[];
+        nohoist?: string[];
+      };
 };
+
+export type ModularPackageJson = Extend<PackageJson, PackageJsonOverrides>;
 
 export function getModularType(dir: string): ModularType | undefined {
   const packageJsonPath = path.join(dir, 'package.json');
@@ -46,17 +65,4 @@ export default function isModularType(dir: string, type: PackageType): boolean {
 
 export function isValidModularType(dir: string): boolean {
   return ModularTypes.includes(getModularType(dir) as ModularType);
-}
-
-export function isValidModularRootPackageJson(dir: string): boolean {
-  const packageJsonPath = path.join(dir, 'package.json');
-  if (fs.existsSync(packageJsonPath)) {
-    const packageJson = fs.readJsonSync(packageJsonPath) as ModularPackageJson;
-    return (
-      packageJson.modular?.type === 'root' &&
-      !!packageJson.private &&
-      !!packageJson?.workspaces?.includes('packages/**')
-    );
-  }
-  return false;
 }
